+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1 / 2016-12-11
+
+* [FIX] Rename `ecs_cluster_total` to `ecs_clusters` metrics
+* [FIX] Rename `ecs_service_total` to `ecs_services` metrics
+
 ## 1.0.0 / 2016-12-11
 
 * [FEATURE] Add `ecs_up` metrics